% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acumulados.R
\name{acumulados}
\alias{acumulados}
\title{ACUMULADOS}
\usage{
acumulados(covid_dataset, variable = NULL)
}
\arguments{
\item{covid_dataset}{Datos provenientes de \code{positividad}, \code{casos},
\code{numero_pruebas}, etc.}

\item{variable}{Variable a acumular}
}
\description{
\code{acumulados} Calcula los casos acumulados
}
\details{
This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Casos acumulados en COAHUILA
datos_covid \%>\%
  casos(entidades = c("COAHUILA")) \%>\%
  acumulados() \%>\%
  covid_plot()

datos_covid \%>\% casos() \%>\% acumulados() \%>\% covid_plot()
}

}
