% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casos_uci.R
\name{casos_uci}
\alias{casos_uci}
\title{CASOS UCI (Unidad Cuidados Intensivos)}
\usage{
casos_uci(
  datos_covid = NULL,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Médica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Síntomas", "Ingreso", "Defunción"),
  tipo_caso = c("Todos", "Sospechosos", "Confirmados COVID",
    "Sospechosos y Confirmados COVID", "Negativos a COVID")
)
}
\arguments{
\item{datos_covid}{If no data is available it automatically downloads COVID-19
information.}

\item{entidades}{Vector con las entidades de las unidades medicas a analizar.
Opciones: \code{AGUASCALIENTES}, \verb{BAJA CALIFORNIA}, \verb{BAJA CALIFORNIA SUR},
\code{CAMPECHE}, \code{CHIAPAS}, \code{CHIHUAHUA}, \verb{CIUDAD DE MEXICO},
\verb{COAHUILA DE ZARAGOZA} , \code{COLIMA}, \code{DURANGO}, \code{GUANAJUATO}, \code{GUERRERO},
\code{HIDALGO}, \code{JALISCO}, \code{MEXICO}, \verb{MICHOACAN DE OCAMPO}, \code{MORELOS},\code{NAYARIT}
\verb{NUEVO LEON}, \code{OAXACA} ,\code{PUEBLA}, \code{QUERETARO},\verb{QUINTANA ROO},
\verb{SAN LUIS POTOSI}, \code{SINALOA}, \code{SONORA}, \code{TABASCO}, \code{TAMAULIPAS},\code{TLAXCALA},
\verb{VERACRUZ DE IGNACIO DE LA LLAVE}, \code{YUCATAN}, \code{ZACATECAS}}

\item{group_by_entidad}{Si junta las entidades en una sola
o bien las muestra por separado sin agrupar.}

\item{entidad_tipo}{Selecciona \verb{Unidad Medica}, \code{Nacimiento} o \code{Residencia}.
por default incluye \verb{Unidad Medica}}

\item{fecha_tipo}{Selecciona \code{Ingreso}, \code{Sintomas} o \code{Defuncion} por default
incluye fecha de \code{Sintomas}}

\item{tipo_caso}{Selecciona \code{Todos}, \verb{Confirmados COVID},\code{Sospechosos},
\verb{Sospechosos y Confirmados COVID}, \verb{Negativos a COVID}}
}
\description{
\code{casos_uci} Calcula el numero de casos registrados por fecha
con TIPO_PACIENTE = HOSPITALIZADO y UCI = "Sí"
}
\details{
Por default calcula el numero de casos con pacientes en UCI véase
\code{casos}, \code{casos_hospitalizados}, \code{casos_ambulatorios}

This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Positividad nacional
datos_covid \%>\% casos_uci()

#Positividad en Jalisco y Colima
datos_covid \%>\% casos_uci(entidades = c("JALISCO","COLIMA")) \%>\%
         covid_plot()

datos_covid \%>\%
    casos_uci(entidades = c("JALISCO","COLIMA"),
                         tipo_caso == "Confirmados COVID") \%>\%
    covid_plot()

datos_covid \%>\% casos_uci(entidades = c("JALISCO","COLIMA"),
                            group_by_entidad = T) \%>\%
         covid_plot()
}

}
