% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pin_get_download_time.R
\name{check_sites}
\alias{check_sites}
\title{Funcion para verificar la existencia de los sitios web de datos de covid}
\usage{
check_sites(covid_data = TRUE, dictionnary = TRUE)
}
\arguments{
\item{covid_data}{(\strong{opcional}) Variable booleana \code{TRUE} si verifica los sitios de
datos y \code{FALSE} si no los verifica}

\item{dictionnary}{(\strong{opcional}) Variable booleana \code{TRUE} si verifica los sitios del
diccionario y \code{FALSE} si no lo verifica}
}
\value{
Devuelve \code{TRUE} si todos los sitios web existen, \code{FALSE}en caso de que no.
}
\description{
La funcion recorre cada uno de los sitios y verifica su existencia.
}
\examples{
# Verificamos que existan los sitios cambiando a TRUE cualquiera:
check_sites(covid_data = FALSE, dictionnary = FALSE)

}
