% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_plot.R
\name{covid_plot}
\alias{covid_plot}
\title{COVID PLOT: Graficacion de las tablas calculadas por el paquete}
\usage{
covid_plot(covid_dataset, scales = "free")
}
\arguments{
\item{covid_dataset}{Datos provenientes de \code{positividad}, \code{casos},
\code{numero_pruebas}, etc}

\item{scales}{Scales parameter for \code{\link[ggplot2]{facet_wrap}}}
}
\description{
\code{covid_plot} Grafica las tablas calculadas por el paquete
}
\details{
This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Positividad nacional
datos_covid \%>\% positividad() \%>\% covid_plot()

#Positividad en Jalisco y Colima
datos_covid \%>\% numero_pruebas(entidades = c("JALISCO","COLIMA")) \%>\%
         covid_plot()
datos_covid \%>\% numero_pruebas(entidades = c("JALISCO","COLIMA"),
                            group_by_entidad = T) \%>\%
         covid_plot()
}

}
