% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunciones_hospitalizados.R
\name{defunciones_hospitalizados}
\alias{defunciones_hospitalizados}
\title{DEFUNCIONES HOSPITALIZADOS}
\usage{
defunciones_hospitalizados(
  datos_covid = NULL,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Médica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Defunción", "Síntomas", "Ingreso"),
  tipo_caso = c("Todos", "Sospechosos", "Confirmados COVID",
    "Sospechosos y Confirmados COVID", "Negativos a COVID")
)
}
\arguments{
\item{datos_covid}{If no data is available it automatically downloads COVID-19
information.}

\item{entidades}{Vector con las entidades de las unidades medicas a analizar.
Opciones: \code{AGUASCALIENTES}, \verb{BAJA CALIFORNIA}, \verb{BAJA CALIFORNIA SUR},
\code{CAMPECHE}, \code{CHIAPAS}, \code{CHIHUAHUA}, \verb{CIUDAD DE MEXICO},
\verb{COAHUILA DE ZARAGOZA} , \code{COLIMA}, \code{DURANGO}, \code{GUANAJUATO}, \code{GUERRERO},
\code{HIDALGO}, \code{JALISCO}, \code{MEXICO}, \verb{MICHOACAN DE OCAMPO}, \code{MORELOS},\code{NAYARIT}
\verb{NUEVO LEON}, \code{OAXACA} ,\code{PUEBLA}, \code{QUERETARO},\verb{QUINTANA ROO},
\verb{SAN LUIS POTOSI}, \code{SINALOA}, \code{SONORA}, \code{TABASCO}, \code{TAMAULIPAS},\code{TLAXCALA},
\verb{VERACRUZ DE IGNACIO DE LA LLAVE}, \code{YUCATAN}, \code{ZACATECAS}}

\item{group_by_entidad}{Si junta las entidades en una sola
o bien las muestra por separado sin agrupar.}

\item{entidad_tipo}{Selecciona \verb{Unidad Medica}, \code{Nacimiento} o \code{Residencia}.
por default incluye \verb{Unidad Medica}}

\item{fecha_tipo}{Selecciona \code{Ingreso}, \code{Sintomas} o \code{Defuncion} por default
incluye fecha de \code{Sintomas}}

\item{tipo_caso}{Selecciona \code{Todos}, \verb{Confirmados COVID},\code{Sospechosos},
\verb{Sospechosos y Confirmados COVID}, \verb{Negativos a COVID}}
}
\description{
\code{defunciones_hospitalizados} Calcula el numero de casos registrados
por fecha de defunción en UCI
}
\details{
Por default calcula el numero de defunciones
vease \code{defunciones},
\code{defunciones_ambulatorios} y \code{defunciones_uci}

This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Defunciones ambulatorios nacional
datos_covid \%>\% defunciones_hospitalizados()

#Positividad en Jalisco y Colima
datos_covid \%>\% defunciones_hospitalizados(entidades = c("JALISCO","COLIMA")) \%>\%
         covid_plot()

datos_covid \%>\%
    defunciones_hospitalizados(entidades = c("JALISCO","COLIMA"),
    tipo_caso == "Confirmados COVID") \%>\%
    acumulados() \%>\%
    covid_plot()

datos_covid \%>\% defunciones_hospitalizados(entidades = c("JALISCO","COLIMA"),
                            group_by_entidad = T) \%>\%
         acumulados() \%>\%
         covid_plot()
}

}
