% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos_ocupacion_hospitalaria.R
\encoding{UTF-8}
\name{descarga_datos_ocupacion_hospitalaria}
\alias{descarga_datos_ocupacion_hospitalaria}
\title{LEE LA BASE DE DATOS DE OCUPACIoN HOSPITALARIA}
\usage{
descarga_datos_ocupacion_hospitalaria(nivel = c("Estatal", "Unidad Médica"))
}
\arguments{
\item{nivel}{"Estatal" o "Unidad Médica"}
}
\value{
\code{data.frame} con los datos de ocupacion hospitalaria
}
\description{
\code{descarga_datos_ocupacion_hospitalaria} Lee los datos de ocupacion hospitalaria de la \verb{RED IRAG}
disponibles en \url{https://www.gits.igg.unam.mx/red-irag-dashboard/reviewHome#}
}
\details{
\emph{This is not an official product / este no es un producto oficial}

\strong{Spanish}
Los datos son descargados de manera automatica en mi Github:
\href{https://github.com/RodrigoZepeda/CapacidadHospitalariaMX}{RodrigoZepeda/CapacidadHospitalariaMX}
y esta funcion los lee de ahí
}
\examples{
\dontrun{
ocupacion_hospitalaria <- descarga_datos_ocupacion_hospitalaria()
}
}
