% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos_red_irag.R
\encoding{UTF-8}
\name{descarga_datos_red_irag}
\alias{descarga_datos_red_irag}
\title{LEE LA BASE DE DATOS DE OCUPACION HOSPITALARIA DE RED IRAG}
\usage{
descarga_datos_red_irag(
  nivel = c("Estatal", "Unidad Médica"),
  cache = NULL,
  use_cache_on_failure = rlang::is_interactive(),
  quiet = TRUE,
  ...
)
}
\arguments{
\item{nivel}{"Estatal" o "Unidad Medica"}

\item{cache}{cache para \code{pins::board_url}.}

\item{use_cache_on_failure}{parametro para \code{pins::board_url}.}

\item{quiet}{booleana para no imprimir mensajes en la consola.}

\item{...}{parametros adicionales para \code{pins::pin_download}.}
}
\value{
\code{data.frame} con los datos de ocupacion hospitalaria
}
\description{
\code{descarga_datos_red_irag} Lee los datos de ocupacion hospitalaria de la \verb{RED IRAG}
disponibles en \url{https://www.gits.igg.unam.mx/red-irag-dashboard/reviewHome#}
}
\details{
Los datos de Red IRAG son descargados diariamente de manera automatica en Github:
\href{https://github.com/RodrigoZepeda/CapacidadHospitalariaMX}{RodrigoZepeda/CapacidadHospitalariaMX}
y esta funcion los lee de ahi
}
\examples{
\dontrun{
#Descarga de datos estatales
ocupacion_hospitalaria <- descarga_datos_red_irag("Estatal")

#Descarga de datos por unidad medica
ocupacion_hospitalaria <- descarga_datos_red_irag("Unidad Medica")
}
}
\seealso{
descarga_datos_variantes_GISAID descarga_datos_abiertos
}
