% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{estima_rt}
\alias{estima_rt}
\title{RT: Número efectivo de reproducción}
\usage{
estima_rt(
  datos_covid,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Medica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Sintomas", "Ingreso", "Defuncion"),
  tipo_clasificacion = c("Sospechosos", "Confirmados COVID", "Negativo a COVID",
    "Inválido", "No realizado"),
  tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO", "NO ESPECIFICADO"),
  list_name = "estima_rt",
  min_date = as.POSIXct("2021/11/21", format = "\%Y/\%m/\%d"),
  max_date = as.POSIXct(Sys.time()),
  method = "parametric_si",
  config = if (requireNamespace("EpiEstim", quietly = TRUE)) {
    
    EpiEstim::make_config(list(mean_si = 3.5, std_si = 1.5))
 } else {
     NULL
 },
  ...
)
}
\arguments{
\item{datos_covid}{(\strong{obligatorio}) Lista de \code{tibble}s resultante de \code{\link[=descarga_datos_abiertos]{descarga_datos_abiertos()}}
o \code{\link[=read_datos_abiertos]{read_datos_abiertos()}}}

\item{entidades}{(\strong{opcional})  Vector con las entidades de las unidades medicas a analizar.
Opciones: \code{AGUASCALIENTES}, \verb{BAJA CALIFORNIA}, \verb{BAJA CALIFORNIA SUR},
\code{CAMPECHE}, \code{CHIAPAS}, \code{CHIHUAHUA}, \verb{CIUDAD DE MEXICO},
\verb{COAHUILA DE ZARAGOZA} , \code{COLIMA}, \code{DURANGO}, \code{GUANAJUATO}, \code{GUERRERO},
\code{HIDALGO}, \code{JALISCO}, \code{MEXICO}, \verb{MICHOACAN DE OCAMPO}, \code{MORELOS},\code{NAYARIT}
\verb{NUEVO LEON}, \code{OAXACA} ,\code{PUEBLA}, \code{QUERETARO},\verb{QUINTANA ROO},
\verb{SAN LUIS POTOSI}, \code{SINALOA}, \code{SONORA}, \code{TABASCO}, \code{TAMAULIPAS},\code{TLAXCALA},
\verb{VERACRUZ DE IGNACIO DE LA LLAVE}, \code{YUCATAN}, \code{ZACATECAS}.}

\item{group_by_entidad}{(\strong{opcional}) \code{TRUE} obtiene los casos para cada entidad reportando
en cada fecha la entidad y los casos en dicha entidad. \code{FALSE}  junta las \code{entidades} sumando
sus casos en una sola observacion por cada fecha.}

\item{entidad_tipo}{(\strong{opcional}) Indica a que se refiere las \code{entidades} seleccionadas. Elige
una de las opciones: \verb{Unidad Medica} (entidad de la unidad medica), \code{Nacimiento}
(entidad de origen del individuo) o \code{Residencia} (entidad donde reside el individuo).}

\item{fecha_tipo}{(\strong{opcional}) Selecciona si la fecha que se utiliza es la fecha de \code{Ingreso}
(si aplica), la fecha de \code{Sintomas} o la de \code{Defuncion} (si aplica). El default es fecha de \code{Sintomas}.}

\item{tipo_clasificacion}{(\strong{opcional})  Vector con el tipo de clasificaciones (por la prueba)
a incluir:\code{Sospechosos},\verb{Confirmados COVID}, \verb{Negativo a COVID}, \verb{Inv\\u00e1lido}, \verb{No realizado}}

\item{tipo_paciente}{(\strong{opcional}) Vector con el tipo de pacientes a incluir. Opciones:
\code{AMBULATORIO}, \code{HOSPITALIZADO}, \verb{NO ESPECIFICADO}. Por default se incluyen todos.}

\item{list_name}{(\strong{opcional}) Asigna un nombre en la lista de datos a la base generada}

\item{min_date}{Mínima fecha a partir de la cual estimar el RT.}

\item{max_date}{Máxima fecha a partir de la cual estimar el RT.}

\item{method}{Metodo para estimar el RT con \code{\link[EpiEstim:estimate_R]{EpiEstim::estimate_R()}}.}

\item{config}{Configuracion para la estimacion del RT usando \code{\link[EpiEstim:make_config]{EpiEstim::make_config()}}.}

\item{...}{Parámetros adicionales para \code{\link[EpiEstim:estimate_R]{EpiEstim::estimate_R()}}.}
}
\value{
Appends a la lista de \code{datos_covid} una nueva entrada de nombre \code{list_name}
(default: \code{casos}) con una base de datos (\code{tibble}) con los
resultados agregados.
\itemize{
\item positividad - Base de datos generara con los datos agregados (el nombre cambia si
se usa \code{list_name}).
\item dict - Diccionario de datos
\item dats - Datos originales (conexion a DB)
\item disconnect  - Funcion para desconectarte de DB
}
}
\description{
\code{estima_rt} Calcula el número efectivo de reproducción por fecha y entidad usando
los metodos de \code{\link[EpiEstim:estimate_R]{EpiEstim::estimate_R()}}. Por default calcula el número efectivo de
reproducción para cada estado.
}
\examples{
\dontrun{
datos_covid <- datosabiertos

# Casos a nivel nacional en los confirmados
datos_covid <- datos_covid |>
  estima_rt(
    tipo_clasificacion = "Confirmados COVID",
    group_by_entidad   = FALSE
  )

# Casos en AGS, CHI en los confirmados
datos_covid |>
  estima_rt(
    entidades = c("CHIHUAHUA", "AGUASCALIENTES"),
    tipo_clasificacion = "Confirmados COVID",
    group_by_entidad = TRUE,
    list_name = "rt_ch_ags"
  ) |>
  plot_covid(
    df_name = "rt_ch_ags", df_date_index = "FECHA_SINTOMAS",
    df_variable = "Mean(R)", df_covariates = "ENTIDAD_FEDERATIVA"
  )
}
}
\seealso{
\code{\link[=descarga_datos_abiertos]{descarga_datos_abiertos()}} \code{\link[=numero_pruebas]{numero_pruebas()}} \code{\link[=cfr]{cfr()}} \code{\link[=chr]{chr()}}
\code{\link[=positividad]{positividad()}} \code{\link[=casos]{casos()}}
}
