% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numero_pruebas_esperando_resultado.R
\name{numero_pruebas_esperando_resultado}
\alias{numero_pruebas_esperando_resultado}
\title{NUMERO DE PRUEBAS PCR ESPERANDO RESULTADO}
\usage{
numero_pruebas_esperando_resultado(
  datos_covid = NULL,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Médica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Síntomas", "Ingreso", "Defunción")
)
}
\arguments{
\item{datos_covid}{If no data is available it automatically downloads COVID-19
information.}

\item{entidades}{Vector con las entidades de las unidades medicas a analizar.
Opciones: \code{AGUASCALIENTES}, \verb{BAJA CALIFORNIA}, \verb{BAJA CALIFORNIA SUR},
\code{CAMPECHE}, \code{CHIAPAS}, \code{CHIHUAHUA}, \verb{CIUDAD DE MEXICO},
\verb{COAHUILA DE ZARAGOZA} , \code{COLIMA}, \code{DURANGO}, \code{GUANAJUATO}, \code{GUERRERO},
\code{HIDALGO}, \code{JALISCO}, \code{MEXICO}, \verb{MICHOACAN DE OCAMPO}, \code{MORELOS},\code{NAYARIT}
\verb{NUEVO LEON}, \code{OAXACA} ,\code{PUEBLA}, \code{QUERETARO},\verb{QUINTANA ROO},
\verb{SAN LUIS POTOSI}, \code{SINALOA}, \code{SONORA}, \code{TABASCO}, \code{TAMAULIPAS},\code{TLAXCALA},
\verb{VERACRUZ DE IGNACIO DE LA LLAVE}, \code{YUCATAN}, \code{ZACATECAS}}

\item{group_by_entidad}{Si junta las entidades en una sola
o bien las muestra por separado sin agrupar.}

\item{entidad_tipo}{Selecciona \verb{Unidad Medica}, \code{Nacimiento} o \code{Residencia}.
por default incluye \verb{Unidad Medica}}

\item{fecha_tipo}{Selecciona \code{Ingreso}, \code{Sintomas} o \code{Defuncion} por default
incluye fecha de \code{Sintomas}}
}
\description{
\code{numero_pruebas_esperando_resultado} Calcula el numero de pruebas que se
realizaron por fecha y que siguen esperando resultado
}
\details{
Por default calcula el numero de pruebas de laboratorio que siguen
esperando resultado por fecha

This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Positividad nacional
datos_covid \%>\% numero_pruebas_esperando_resultado()

#Positividad en Jalisco y Colima
datos_covid \%>\%
  numero_pruebas_esperando_resultado(entidades = c("JALISCO","COLIMA")) \%>\%
  covid_plot()
datos_covid \%>\%
   numero_pruebas_esperando_resultado(entidades = c("JALISCO","COLIMA"),
                            group_by_entidad = T) \%>\%
   covid_plot()
}

}
