% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positividad_antigenos.R
\name{positividad_antigenos}
\alias{positividad_antigenos}
\title{POSITIVIDAD POR ANTIGENOS}
\usage{
positividad_antigenos(
  datos_covid = NULL,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Médica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Síntomas", "Ingreso", "Defunción")
)
}
\arguments{
\item{datos_covid}{If no data is available it automatically downloads COVID-19
information.}

\item{entidades}{Vector con las entidades de las unidades medicas a analizar.
Opciones: \code{AGUASCALIENTES}, \verb{BAJA CALIFORNIA}, \verb{BAJA CALIFORNIA SUR},
\code{CAMPECHE}, \code{CHIAPAS}, \code{CHIHUAHUA}, \verb{CIUDAD DE MEXICO},
\verb{COAHUILA DE ZARAGOZA} , \code{COLIMA}, \code{DURANGO}, \code{GUANAJUATO}, \code{GUERRERO},
\code{HIDALGO}, \code{JALISCO}, \code{MEXICO}, \verb{MICHOACAN DE OCAMPO}, \code{MORELOS},\code{NAYARIT}
\verb{NUEVO LEON}, \code{OAXACA} ,\code{PUEBLA}, \code{QUERETARO},\verb{QUINTANA ROO},
\verb{SAN LUIS POTOSI}, \code{SINALOA}, \code{SONORA}, \code{TABASCO}, \code{TAMAULIPAS},\code{TLAXCALA},
\verb{VERACRUZ DE IGNACIO DE LA LLAVE}, \code{YUCATAN}, \code{ZACATECAS}}

\item{group_by_entidad}{Si junta las entidades en una sola
o bien las muestra por separado sin agrupar.}

\item{entidad_tipo}{Selecciona \verb{Unidad Medica}, \code{Nacimiento} o \code{Residencia}.
por default incluye \verb{Unidad Medica}}

\item{fecha_tipo}{Selecciona \code{Ingreso}, \code{Sintomas} o \code{Defuncion} por default
incluye fecha de \code{Sintomas}}
}
\description{
\code{positividad_antigenos} Calcula la positividad definida como el radio de
\deqn{\frac{Pruebas positivas}{Total de pruebas}} para las pruebas rapidas
}
\details{
Por default calcula la positividad total en pruebas rapidas (antigenos)
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Positividad nacional por PCR
datos_covid \%>\% positividad_antigenos()

#Positividad en Jalisco y Colima
datos_covid \%>\% positividad_antigenos(entidades = c("JALISCO","COLIMA")) \%>\%
       covid_plot()
datos_covid \%>\% positividad_antigenos(entidades = c("JALISCO","COLIMA"),
                                group_by_entidad = T) \%>\%
       covid_plot()
}

}
\seealso{
\code{positividad_pcr} y \code{positividad}
This is not an official product / este no es un producto oficial
}
