% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datos_abiertos.R
\encoding{UTF-8}
\name{read_datos_abiertos}
\alias{read_datos_abiertos}
\title{LEE LA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA QUE YA DESCARGASTE}
\usage{
read_datos_abiertos(
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  dict_file = "diccionario_covid.RData",
  download_dict = FALSE,
  download_method = "curl",
  save_dict = download_dict,
  file_download_dictionary = tempfile(),
  remove_zip_after_download = TRUE,
  site.covid.dic = paste0("http://datosabiertos.salud.gob.mx/gobmx/salud/",
    "datos_abiertos/diccionario_datos_covid19.zip")
)
}
\arguments{
\item{user}{User for \code{dbConnect} i.e. your \code{MariaDB} user}

\item{password}{password for \code{dbConnect} i.e. your \code{MariaDB} password}

\item{dbname}{Database name for \code{dbConnect} i.e. what database is going to be stored}

\item{host}{Host for \code{dbConnect} i.e. your \code{MariaDB} host (usually localhost)}

\item{group}{Group for \code{dbConnect} i.e. your \code{MariaDB} group (can be NULL)}

\item{port}{Port connection for \code{MariaDB}}

\item{tblname}{Name of table to save in \code{MariaDB}}

\item{dict_file}{RData File where to save/read the dictionary.}

\item{download_dict}{If download the dictionary from \code{site.covid.dic}. If
\code{download_dict} is \code{FALSE} then \code{dict_file} must be specified to read the dictionary.}

\item{download_method}{Methods for download file (default = "curl"). Other
options are "internal", "wininet" (Windows) "libcurl", "wget", "curl". See
\code{\link[utils]{download.file}} / Metodos para descargar el archivo
(default = "curl"). Otras opciones incluyen "interno","wininet" (Windows) "
libcurl", "wget", "curl". Vease \code{\link[utils]{download.file}}.}

\item{save_dict}{If the downloaded dictionary is to be saved in file \code{dict_file}}

\item{file_download_dictionary}{Name of file to save the dictionary /
Nombre del archivo donde guardar el diccioanrio.}

\item{remove_zip_after_download}{If the downloaded zip file should be saved /
Si los archivos zip descargados deben ser almacenados.}

\item{site.covid.dic}{Site for download the covid dictionary for the data}
}
\value{
List of values:
\itemize{
\item dats - Database table (if MARIADB) or database in tibble (if tibble)
\item con - Database connection (if MARIADB) or \code{NULL} (if tibble)
\item dict - List of tibbles containing the whole dictionary
}
}
\description{
\code{read_datos_abiertos} Lee los datos abiertos almacenados en tu base de \code{MariaDB} que
bajaste con \code{descarga_datos_abiertos}
}
\details{
\emph{This is not an official product / este no es un producto oficial}

\strong{Spanish}
Necesitas tener una instalación funcionando de \href{https://mariadb.com/}{\code{MariaDB}}. El programa
se encarga de leer la base de datos de la DGE que bajamos con \code{descarga_datos_abiertos}.
Basta con leer la tabla de nombre \code{covidmx} dentro de tu database \code{dbname}.
Asegúrate de que tu usuario \code{user} tenga permisos de lectura.
Para más información sobre instalación y uso de \code{MariaDB} consulta la viñeta \strong{rellenar}.
}
\examples{
\dontrun{
datos_covid <- read_datos_abiertos()
}
}
