% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Suavizamiento de los casos}
\usage{
smooth(
  covid_dataset,
  variable = NULL,
  method = c("splines", "kernel", "lowess", "rollmean"),
  df = 10,
  bw = 100,
  f = 0.1,
  k = 7
)
}
\arguments{
\item{covid_dataset}{Datos provenientes de \code{positividad}, \code{casos},
\code{numero_pruebas}, etc.}

\item{variable}{Variable a suavizar.}

\item{method}{\code{splines}, \code{lowess}, \code{kernel}, \code{rollmean}. Metodo de suavizamiento}

\item{df}{Degrees of freedom for spline smoothing}

\item{bw}{Banwidth for kernel}

\item{f}{Parameter for lowess}

\item{k}{Days to consider in rolling mean}
}
\description{
\code{smooth} Suaviza los casos reportados
}
\details{
Por ahora solo esta la opcion de splines, lowess y kernel.

This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")


#Positividad en Jalisco y Colima
datos_covid \%>\% casos(entidades = c("JALISCO","COLIMA")) \%>\%
         smooth() \%>\%
         covid_plot()
datos_covid \%>\% casos(entidades = c("JALISCO","COLIMA")) \%>\%
         smooth(method = "kernel") \%>\%
         covid_plot()
datos_covid \%>\% casos(entidades = c("JALISCO","COLIMA")) \%>\%
         smooth(method = "lowess", f = 0.1) \%>\%
         covid_plot()
datos_covid \%>\% casos(entidades = c("JALISCO","COLIMA")) \%>\%
         smooth(method = "rollmean", k = 14) \%>\%
         covid_plot()
}

}
