% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasa_cambio.R
\name{tasa_cambio}
\alias{tasa_cambio}
\title{TASA DE CAMBIO}
\usage{
tasa_cambio(covid_dataset, variable = NULL, quiet = FALSE)
}
\arguments{
\item{covid_dataset}{Datos provenientes de \code{positividad}, \code{casos},
\code{numero_pruebas}, etc.}

\item{variable}{Variable sobre la cual calcular la tasa de cambio}

\item{quiet}{Show warning?}
}
\description{
\code{tasa_cambio} Calcula la diferencia de casos del dia t y el t-1
}
\details{
This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")

#Positividad nacional
datos_covid \%>\% casos() \%>\% tasa_cambio() \%>\% covid_plot()

}

}
