% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidmx-package.R
\docType{package}
\name{covidmx-package}
\alias{covidmx}
\alias{covidmx-package}
\title{covidmx: Descarga y analiza datos de COVID-19 en México}
\description{
Herramientas para el análisis de datos de COVID-19 en México. Descarga y analiza los datos para COVID-19 de la Direccion General de Epidemiología de México (DGE) \url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}, la Red de Infecciones Respiratorias Agudas Graves (Red IRAG) \url{https://www.gits.igg.unam.mx/red-irag-dashboard/reviewHome} y la Iniciativa Global para compartir todos los datos de influenza (GISAID) \url{https://gisaid.org/}. English: Downloads and analyzes data of COVID-19 from the Mexican General Directorate of Epidemiology (DGE), the Network of Severe Acute Respiratory Infections (IRAG network),and the Global Initiative on Sharing All Influenza Data GISAID.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RodrigoZepeda/covidmx}
  \item \url{https://rodrigozepeda.github.io/covidmx/}
  \item Report bugs at \url{https://github.com/RodrigoZepeda/covidmx/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com} (\href{https://orcid.org/0000-0003-4471-5270}{ORCID})

Authors:
\itemize{
  \item Mauricio Hernandez-Avila
}

Other contributors:
\itemize{
  \item Alberto Almuiña (Author of included zzz fragment) [contributor]
  \item Jonah Gabry (Author of included cmdstanr fragment) [contributor]
  \item Rok Češnovar (Author of included cmdstanr fragment) [contributor]
  \item Ben Bales (Author of included cmdstanr fragment) [contributor]
  \item Mitzi Morris (Author of included cmdstanr fragment) [contributor]
  \item Mikhail Popov (Author of included cmdstanr fragment) [contributor]
  \item Mike Lawrence (Author of included cmdstanr fragment) [contributor]
  \item William Michael Landau (Author of included cmdstanr fragment) [contributor]
  \item Jacob Socolar (Author of included cmdstanr fragment) [contributor]
  \item Andrew Johnson (Author of included cmdstanr fragment) [contributor]
  \item Instituto Mexicano del Seguro Social [copyright holder, funder]
}

}
\keyword{internal}
