% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidmx-package.R
\docType{package}
\name{covidmx-package}
\alias{covidmx}
\alias{covidmx-package}
\title{covidmx: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
