% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos_abiertos.R
\encoding{UTF-8}
\name{descarga_datos_abiertos}
\alias{descarga_datos_abiertos}
\title{DESCARGA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA}
\usage{
descarga_datos_abiertos(
  download_method = "curl",
  file_download_data = tempfile(),
  file_download_dictionary = tempfile(),
  remove_zip_after_download = TRUE,
  quiet = FALSE,
  parse_warnings = FALSE,
  language = c("English", "Español"),
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  dict_file = "diccionario_covid.RData",
  download_dict = TRUE,
  save_dict = download_dict,
  site.covid = paste0("http://datosabiertos.salud.gob.mx/gobmx/salud",
    "/datos_abiertos/datos_abiertos_covid19.zip"),
  site.covid.dic = paste0("http://datosabiertos.salud.gob.mx/gobmx/salud/",
    "datos_abiertos/diccionario_datos_covid19.zip"),
  nthreads = max(parallel::detectCores() - 1, 1),
  unzip_command = ifelse(tolower(.Platform$OS.type) == "windows",
    "\\"C:\\\\Program Files\\\\7-Zip\\\\7z.exe\\"", "unzip"),
  unzip_args = ifelse(tolower(.Platform$OS.type) == "windows", "-x", "-o"),
  check_unzip_install = TRUE,
  ...
)
}
\arguments{
\item{download_method}{Methods for download file (default = "curl"). Other
options are "internal", "wininet" (Windows) "libcurl", "wget", "curl". See
\code{\link[utils]{download.file}} / Metodos para descargar el archivo
(default = "curl"). Otras opciones incluyen "interno","wininet" (Windows) "
libcurl", "wget", "curl". Vease \code{\link[utils]{download.file}}.}

\item{file_download_data}{Name of file to save the data / Nombre del archivo
para guardar los datos.}

\item{file_download_dictionary}{Name of file to save the dictionary /
Nombre del archivo donde guardar el diccioanrio.}

\item{remove_zip_after_download}{If the downloaded zip file should be saved /
Si los archivos zip descargados deben ser almacenados.}

\item{quiet}{Show messages? / Mostrar mensajes?}

\item{parse_warnings}{If show parsing errors when reading csv /
Si muestra errores cuando lee el csv}

\item{language}{('Español', 'English') Message languages/ Idiomas del mensaje.}

\item{read_format}{\code{"MariaDB"} or \code{"tibble"} establishes how the database
should be stored. In most machines \code{"tibble"} will result in a memory error}

\item{user}{User for \code{dbConnect} i.e. your \code{MariaDB} user}

\item{password}{password for \code{dbConnect} i.e. your \code{MariaDB} password}

\item{dbname}{Database name for \code{dbConnect} i.e. what database is going to be stored}

\item{host}{Host for \code{dbConnect} i.e. your \code{MariaDB} host (usually localhost)}

\item{group}{Group for \code{dbConnect} i.e. your \code{MariaDB} group (can be NULL)}

\item{port}{Port connection for \code{MariaDB}}

\item{tblname}{Name of table to save in \code{MariaDB}}

\item{dict_file}{RData File where to save/read the dictionary.}

\item{download_dict}{If download the dictionary from \code{site.covid.dic}. If
\code{download_dict} is \code{FALSE} then \code{dict_file} must be specified to read the dictionary.}

\item{save_dict}{If the downloaded dictionary is to be saved in file \code{dict_file}}

\item{site.covid}{Site for download the covid information}

\item{site.covid.dic}{Site for download the covid dictionary for the data}

\item{nthreads}{Number of threads for writing to \code{MariaDB}.}

\item{unzip_command}{Forma de extraer la base de datos de datos abiertos. La forma de
llamarla es con \code{system2(unzip_command, args = c(unzip_args, file_download_data))}.}

\item{unzip_args}{Argumentos de extraccion de la base de datos de datos abiertos. La forma de
llamarla es con \code{system2(unzip_command, args = c(unzip_args, file_download_data))}.}

\item{check_unzip_install}{Bandera de verificacion para checar si tienes lo necesario para
unzippear los datos.}

\item{...}{Parametros adicionales para \code{DBI::dbConnect}.}
}
\value{
List of values:
\itemize{
\item dats - Database table (if MARIADB) or database in tibble (if tibble)
\item con - Database connection (if MARIADB) or \code{NULL} (if tibble)
\item dict - List of tibbles containing the whole dictionary
}
}
\description{
\code{descarga_datos_abiertos} downloads and labels
all COVID-19 data for Mexico from the Direccion General de EpidemiologIa at
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}.

\code{descarga_datos_abiertos} descarga y etiqueta todos los datos de COVID-19
de Mexico de la Direccion General de EpidemiologIa en
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}.
}
\details{
\emph{This is not an official product / este no es un producto oficial}

\strong{English}
While reading the data it is perfectly normal to get a \code{Warning} for
\code{parsing}: the dataset has issues.

\strong{Spanish}
En la lectura de datos es normal tener un \code{Warning} por \code{parsing}:
la base tiene errores de registro.

Necesitas tener una instalacion funcionando de \href{https://mariadb.com/}{\code{MariaDB}}. El programa
se encarga de descargar la base de datos de la DGE, abrir el archivo \code{.zip} y crear una
tabla de nombre \code{covidmx} dentro de tu database \code{dbname}. Asegurate de que tu usuario \code{user}
tenga permisos de escritura. Para mas informacion sobre instalacion y uso de \code{MariaDB}
consulta \href{https://rodrigozepeda.github.io/covidmx/articles/Instalacion_de_MARIADB.html}{el articulo correspondiente}.

Si tienes RAM que te sobre puedes no crear una base de datos en \code{MariaDB} sino leer directo
el archivo \code{csv}. Esto se logra con \code{read_format = tibble}. No lo recomiendo pues puedes
terminar con tu sesion de \code{R} si se te acaba la memoria.

\emph{Windows} Para abrir el archivo \code{.zip} requieres tambien descargar e instalar \href{https://www.7-zip.org/}{\verb{7Zip}}
por default el sistema lo busca en \verb{C:\\\\Program Files\\\\7-Zip\\\\7z.exe} pero si no esta ese
directorio es necesario que en \code{unzip_command} especifiques el camino donde se instalo \verb{7z.exe}.
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")
}
}
