% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos_abiertos.R
\name{descarga_datos_abiertos}
\alias{descarga_datos_abiertos}
\title{DESCARGA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA}
\usage{
descarga_datos_abiertos(
  download_method = "curl",
  file_download_data = tempfile(),
  file_download_dictionary = tempfile(),
  remove_zip_after_download = TRUE,
  quiet = FALSE,
  parse_dictionary = TRUE,
  parse_warnings = FALSE,
  language = "English"
)
}
\arguments{
\item{download_method}{Methods for download file (default = "curl"). Other
options are "internal", "wininet" (Windows) "libcurl", "wget", "curl". See
\code{\link[utils]{download.file}} / Metodos para descargar el archivo
(default = "curl"). Otras opciones incluyen "interno","wininet" (Windows) "
libcurl", "wget", "curl". Vease \code{\link[utils]{download.file}}.}

\item{file_download_data}{Name of file to save the data / Nombre del archivo
para guardar los datos.}

\item{file_download_dictionary}{Name of file to save the dictionary /
Nombre del archivo donde guardar el diccioanrio.}

\item{remove_zip_after_download}{If the downloaded zip file should be saved /
Si los archivos zip descargados deben ser almacenados.}

\item{quiet}{Show messages? / Mostrar mensajes?}

\item{parse_dictionary}{Change labels to values of dictionary? Default = TRUE
else the data is downloaded and presented as is. / Se agregan etiquetas
a partir del diccionario? Default = TRUE en caso contrario solo se presentan los
datos descargados como estan.}

\item{parse_warnings}{If show parsing errors when reading csv /
Si muestra errores cuando lee el csv}

\item{language}{('Espanol', 'English') Message languages/ Idiomas del mensaje.}
}
\description{
\code{descarga_datos_abiertos} downloads and labels
all COVID-19 data for Mexico from the Direccion General de EpidemiologIa at
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}.

\code{descarga_datos_abiertos} descarga y etiqueta todos los datos de COVID-19
de Mexico de la Direccion General de EpidemiologIa en
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}.
}
\details{
While reading the data it is perfectly normal to get a \code{Warning} for
\code{parsing}: the dataset has issues.
En la lectura de datos es normal tener un \code{Warning} por \code{parsing}:
la base viene chueca.

This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")
}

}
