% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga.R
\encoding{UTF-8}
\name{descarga_datos_abiertos}
\alias{descarga_datos_abiertos}
\alias{descarga_db}
\alias{descarga_diccionario}
\alias{descarga_db_datos_abiertos_tbl}
\alias{descarga_db_diccionario_ssa}
\alias{unzip_db_datos_abiertos_tbl}
\alias{unzip_db_diccionario_ssa}
\alias{parse_db_diccionario_ssa}
\alias{parse_db_datos_abiertos_tbl}
\alias{pega_db_datos_abiertos_tbl_y_diccionario}
\title{Descarga de datos abiertos}
\usage{
descarga_datos_abiertos(
  read_format = c("MariaDB", "tibble"),
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid = paste0("http://datosabiertos.salud.gob.mx",
    "/gobmx/salud/datos_abiertos/datos", "_abiertos_covid19.zip"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  check_unzip_install = TRUE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  use_dict = TRUE,
  datos_abiertos_zip_path = NULL,
  datos_abiertos_unzipped_path = NULL,
  datos_abiertos_tbl = NULL,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  quiet = FALSE,
  cache_datos = NULL,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name = "datos_abiertos",
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_datos_abiertos_tbl_args = list(),
  descarga_db_diccionario_ssa_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

descarga_db(
  read_format = c("MariaDB", "tibble"),
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid = paste0("http://datosabiertos.salud.gob.mx",
    "/gobmx/salud/datos_abiertos/datos", "_abiertos_covid19.zip"),
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  check_unzip_install = TRUE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  datos_abiertos_zip_path = NULL,
  datos_abiertos_unzipped_path = NULL,
  datos_abiertos_tbl = NULL,
  quiet = FALSE,
  board_url_name = "datos_abiertos",
  cache = NULL,
  use_cache_on_failure = TRUE,
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_datos_abiertos_tbl_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

descarga_diccionario(
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  quiet = FALSE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  board_url_name_dict = "diccionario_covid",
  cache_diccionario = NULL,
  use_cache_on_failure = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  download_file_args_dict = list(method = "curl", destfile = tempfile(), quiet = quiet),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  descarga_db_diccionario_ssa_args = list()
)

descarga_db_datos_abiertos_tbl(
  download_process = c("pins", "download.file"),
  site.covid = paste0("http://datosabiertos.salud.gob.mx",
    "/gobmx/salud/datos_abiertos/datos", "_abiertos_covid19.zip"),
  quiet = FALSE,
  board_url_name = "datos_abiertos",
  cache = NULL,
  use_cache_on_failure = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  ...
)

descarga_db_diccionario_ssa(
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  quiet = FALSE,
  board_url_name_dict = "diccionario_covid",
  cache_diccionario = NULL,
  use_cache_on_failure = TRUE,
  force_download = FALSE,
  show_warnings = TRUE,
  download_file_args_dict = list(method = "curl", destfile = tempfile(), quiet = quiet),
  ...
)

unzip_db_datos_abiertos_tbl(
  datos_abiertos_zip_path,
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  check_unzip_install = TRUE,
  quiet = FALSE,
  clear_zip = FALSE
)

unzip_db_diccionario_ssa(
  diccionario_zip_path,
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  clear_zip = FALSE
)

parse_db_diccionario_ssa(diccionario_unzipped_path, clear_csv = FALSE)

parse_db_datos_abiertos_tbl(
  datos_abiertos_unzipped_path,
  read_format = c("MariaDB", "tibble"),
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  quiet = TRUE,
  clear_csv = FALSE,
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

pega_db_datos_abiertos_tbl_y_diccionario(datos_abiertos_tbl, diccionario)
}
\arguments{
\item{read_format}{\code{"MariaDB"} o \code{"tibble"} establece el formato
de lectura de la base de datos. En la mayoria de los casos \code{"tibble"} va a
resultar en un error de memoria.}

\item{driver}{An \code{SQL} driver for \code{dbConnect}}

\item{sqlimport}{An import command similar to \code{mysqlimport}}

\item{user}{Usuario para \code{dbConnect} i.e. tu usuario de \code{MariaDB}.
Puedes ponerlo como variable ambiental con \code{Sys.setenv("MariaDB_user" = "tu_usuario")}}

\item{password}{password para \code{dbConnect} i.e. tu password de \code{MariaDB}
Puedes ponerlo como variable ambiental con \code{Sys.setenv("MariaDB_password" = "tu_passwodrd")}}

\item{dbname}{Nombre de la base de datos \code{DATABASE} para\code{dbConnect} puedes ponerlo
con \code{Sys.setenv("MariaDB_dbname" = "covidmx")}}

\item{host}{(\strong{opcional}) Host para \code{dbConnect}. Si no sabes cual tienes seguro es \code{localhost}. Puedes
ponerlo con \code{Sys.setenv("MariaDB_host" = "localhost")}}

\item{group}{(\strong{opcional})  Grupo de \code{dbConnect} i.e. tu grupo de \code{MariaDB}. Ponlo como variable ambiental
usando \code{Sys.setenv("MariaDB_group" = "tu_grupo")}}

\item{port}{Puerto de la conexion con \code{MariaDB}. Puedes ponerlo con
\code{Sys.setenv("MariaDB_group" = "tu_grupo")}}

\item{tblname}{Nombre de la tabla dentro de el \code{DATABASE} \code{dbname} donde guardar los datos
por default se llama \code{covidmx}.}

\item{nthreads}{Número de \code{threads} para escribir el archivo en \code{MariaDB}.}

\item{download_process}{(\strong{opcional})  Metodo para descargar ya sea \code{pins} o \code{download.file}.
Se recomienda \code{pins} pues guarda en memoria la fecha de la ultima descarga y analiza
si ha pasado mas de un dia desde la descarga. En caso afirmativo verifica si el
archivo ha cambiado y si hubo cambios entonces lo descarga.}

\item{site.covid}{Sitio web con el vinculo al archivo \code{.zip} de los datos abiertos. Puedes
cambiarlo por uno de los historicos, por ejemplo.}

\item{site.covid.dic}{Sitio para descarga del diccionario de datos.}

\item{unzip_command}{Forma de extraer la base de datos de datos abiertos. La forma de
llamarla es con \code{system2(unzip_command, args = c(unzip_args, file_download_data))}.}

\item{unzip_args}{Argumentos de extraccion de la base de datos de datos abiertos. La forma de
llamarla es con \code{system2(unzip_command, args = c(unzip_args, file_download_data))}.}

\item{unzip_args_dict}{Lista de argumentos para usar \code{utils::unzip} en el diccionario de datos.}

\item{check_unzip_install}{Bandera de verificacion para checar si tienes lo necesario para
unzippear los datos.}

\item{clear_zip}{Si borrar los archivos \code{.zip} descargados para el diccionario y los datos
abiertos. No se recomienda si estas usando \code{pins}. Ve la nota para mas informacion.}

\item{clear_csv}{Si borrar los archivos \code{.csv} que se generan despues de abrir el zip. El
default es que si.}

\item{use_dict}{Si descargar el diccionario de \code{site.covid.dic}.}

\item{datos_abiertos_zip_path}{Camino a los datos abiertos si ya los descargaste en \code{zip}}

\item{datos_abiertos_unzipped_path}{Camino a los datos abiertos \code{csv} si ya los descargaste y
descomprimiste el archivo \code{zip} en un \code{csv}}

\item{datos_abiertos_tbl}{Lo que resulta de realizar una descarga de los datos abiertos
usando \code{descarga_db}}

\item{diccionario_zip_path}{Camino al diccionario si ya losdescargaste en \code{zip}}

\item{diccionario_unzipped_path}{Camino al diccionario \code{csv} si ya lo descargaste y
descomprimiste el archivo \code{zip} en un \code{csv}}

\item{diccionario}{Lo que resulta de realizar una descarga del diccionario
usando \code{descarga_diccionario}}

\item{quiet}{(\strong{opcional}) Variable para no mostrar mensajes}

\item{cache_datos}{Direccion donde guardar los datos en memoria usando \code{pins} para no tener
que volver a descargarlos si nada ha cambiado}

\item{use_cache_on_failure}{Booleana. Establece que si no se pueden descargar datos nuevos
utilice los que tenga en memoria.}

\item{cache_diccionario}{Direccion donde guardar el diccionario en memoria usando \code{pins} para no tener
que volver a descargarlo si nada ha cambiado}

\item{force_download}{analiza si cambio el pin y descarga datos nuevos en caso afirmativo}

\item{show_warnings}{si arrojar \code{warnings}}

\item{board_url_name}{Establece el nombre del \code{pins::board_url} para los datos abiertos}

\item{board_url_name_dict}{Establece el nombre del \code{pins::board_url} para los datos abiertos}

\item{download_file_args}{Lista de argumentos adicionales para \code{download.file} de los datos
si se elige este metodo para descargar.}

\item{descarga_db_datos_abiertos_tbl_args}{Lista con argumentos adicionales para el
\code{pins::pin_download} de datos abiertos}

\item{descarga_db_diccionario_ssa_args}{Lista con argumentos adicionales para el
\code{pins::pin_download} de datos abiertos}

\item{prestatement}{Opciones adicionales para \code{MariaDB} yo pongo las que me han funcionado
en que sea mas veloz la carta.}

\item{additional_mysqlimport_flags}{La instalacion se hace con el siguiente comando para \code{MariaDB}:
\preformatted{
mysqlimport --default-character-set=UTF8 --fields-terminated-by=',' --ignore-lines=1
--fields-enclosed-by='\"' --lines-terminated-by='\\n' --user={user}" --password={password}
--use-threads={nthreads} {silent_flag} {additional_mysqlimport_flags} {dbname} ./{tblname}.csv
}}

\item{load_data_infile_options}{Lista de opciones por si la importacion con \code{mysqlimport} no
funciona se use un \verb{LOAD DATA INFILE} como aquí \url{https://dev.mysql.com/doc/refman/8.0/en/load-data.html}
Las variables se pasan como lista asi por ejemplo si queremos agregar la variable local se
pone \code{list("LOCAL" = "LOCAL")} y si no la queremos \code{list("LOCAL" = "")}.}

\item{...}{Parametros adicionales para \code{pins::pin_download}}

\item{cache}{parametro para el cache de \code{pins::board_url}}

\item{download_file_args_dict}{Lista de argumentos adicionales para \code{download.file} del diccionario
si se elige este metodo de descarga.}
}
\value{
List of values:
\itemize{
\item dats        - Tabla conectada mediante \code{DBI::dbConnect} (si \code{MariaDB}) o
tibble (si \code{tibble})
\item disconnect  - Funcion para cerrar la conexion a la base de datos.
\item dict        - Lista de \code{tibble}s con el diccionario de datos para cada variable
}
}
\description{
Conjunto de funciones para la descarga de datos abiertos de
la Direccion General de Epidemiologia

La funcion de descarga principal es \code{descarga_datos_abiertos} llama las siguientes funciones
en orden:
\itemize{
\item \code{\link[=descarga_diccionario]{descarga_diccionario()}}
\item \code{\link[=descarga_db]{descarga_db()}}
\item \code{\link[=pega_db_datos_abiertos_tbl_y_diccionario]{pega_db_datos_abiertos_tbl_y_diccionario()}}
}

A su vez \code{descarga_diccionario} ejecuta las siguientes para obtener el diccionario de datos:
\itemize{
\item \code{\link[=descarga_db_diccionario_ssa]{descarga_db_diccionario_ssa()}}
\item \code{\link[=unzip_db_diccionario_ssa]{unzip_db_diccionario_ssa()}}
\item \code{\link[=parse_db_diccionario_ssa]{parse_db_diccionario_ssa()}}
}

A su vez \code{descarga_db} ejecuta las siguientes para obtener los datos abiertos:
\itemize{
\item \code{\link[=descarga_db_datos_abiertos_tbl]{descarga_db_datos_abiertos_tbl()}}
\item \code{\link[=unzip_db_datos_abiertos_tbl]{unzip_db_datos_abiertos_tbl()}}
\item \code{\link[=parse_db_datos_abiertos_tbl]{parse_db_datos_abiertos_tbl()}}
}

Si en algun momento se interrumpio la descarga o hubo problemas de conexion o detuviste
el proceso de generacion de la base de datos abiertos puedes llamar a las funciones
de \code{\link[=read_datos_abiertos]{read_datos_abiertos()}}
}
\details{
Necesitas tener una instalacion funcionando de \href{https://mariadb.com/}{\code{MariaDB}}. El programa
se encarga de descargar la base de datos de la DGE, abrir el archivo \code{.zip} y crear una
tabla de nombre \code{covidmx} dentro de tu database \code{dbname}. Asegurate de que tu usuario \code{user}
tenga permisos de escritura. Para mas informacion sobre instalacion y uso de \code{MariaDB}
consulta \href{https://rodrigozepeda.github.io/covidmx/articles/Instalacion_de_MARIADB.html}{el articulo correspondiente}.

Si tienes RAM que te sobre puedes no crear una base de datos en \code{MariaDB} sino leer directo
el archivo \code{csv}. Esto se logra con \code{read_format = tibble}. No lo recomiendo pues puedes
terminar con tu sesion de \code{R} si se te acaba la memoria.

\emph{Windows} Para abrir el archivo \code{.zip} requieres tambien descargar e instalar \href{https://www.7-zip.org/}{\verb{7Zip}}
por default el sistema lo busca en \verb{C:\\\\Program Files\\\\7-Zip\\\\7z.exe} pero si no esta ese
directorio es necesario que en \code{unzip_command} especifiques el camino donde se instalo \verb{7z.exe}.
}
\note{
No te recomiendo borrar el cache con \code{clear_zip} o editarlo por cualquier otro medio si
estas usando \code{pins} pues puede romperse la dependencia. Si accidentalmente lo borraste
usa \code{pins::board_cache_path()} para ir al \code{path} y borrar manualmente toda la carpeta.
}
\section{Uso de \code{pins}}{


Para almacenar los datos se utiliza un pequenio cambio sobre la libreria \code{pins}. Los datos
se descargan y se almacenan en cache junto con informacion sobre cuando fue la descarga. Si
no ha pasado un dia desde la ultima descarga no se descarga nada nuevo. Si los datos que
se tienen no han cambiado respecto a lo que esta en linea tampoco se vuelven a descargar aunque
haya pasado mas de un dia.

\strong{Si se te fue el Internet} No te preocupes, \code{pins} lee tu descarga mas reciente.

Para ver donde estan descargados tus datos usa \code{pins::board_cache_path()}. Para borrarlos usa
\code{pins::cache_prune()}.
}

\examples{
\dontrun{
# Descarga de la base de datos junto con diccionario en MariaDB
datos_covid <- descarga_datos_abiertos()

# Luego haces algo con esos datos...

# Cuando terminas cierras la sesión:
datos_covid$disconnect()

# Descarga solo el diccionario
diccionario <- descarga_diccionario()

# O bien descarga solo los datos abiertos
datos_abiertos <- descarga_db()

# Pegalos en el formato que se necesita para el resto de funciones
datos_covid <- pega_db_datos_abiertos(datos_abiertos, diccionario)

# Tambien puedes descargar paso por paso
datos_abiertos <- descarga_db_datos_abiertos_tbl() \%>\% # Descarga
  unzip_db_datos_abiertos_tbl() \%>\% # Unzippea
  parse_db_datos_abiertos_tbl() # MariaDB

# O bien el diccionario
diccionario <- descarga_db_diccionario_ssa() \%>\% # Descarga
  unzip_db_diccionario_ssa() \%>\% # Unzippea
  parse_db_diccionario_ssa() # Tibble
}
}
\seealso{
\link{descarga_datos_red_irag} \link{descarga_datos_variantes_GISAID} \link{read_datos_abiertos} \link{casos}
}
