% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos_abiertos.R
\encoding{UTF-8}
\name{descarga_datos_abiertos}
\alias{descarga_datos_abiertos}
\title{DESCARGA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA}
\usage{
descarga_datos_abiertos(
  download_method = "curl",
  file_download_data = tempfile(),
  file_download_dictionary = tempfile(),
  remove_zip_after_download = TRUE,
  quiet = FALSE,
  parse_warnings = FALSE,
  language = c("English", "Español"),
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1)
)
}
\arguments{
\item{download_method}{Methods for download file (default = "curl"). Other
options are "internal", "wininet" (Windows) "libcurl", "wget", "curl". See
\code{\link[utils]{download.file}} / Metodos para descargar el archivo
(default = "curl"). Otras opciones incluyen "interno","wininet" (Windows) "
libcurl", "wget", "curl". Vease \code{\link[utils]{download.file}}.}

\item{file_download_data}{Name of file to save the data / Nombre del archivo
para guardar los datos.}

\item{file_download_dictionary}{Name of file to save the dictionary /
Nombre del archivo donde guardar el diccioanrio.}

\item{remove_zip_after_download}{If the downloaded zip file should be saved /
Si los archivos zip descargados deben ser almacenados.}

\item{quiet}{Show messages? / Mostrar mensajes?}

\item{parse_warnings}{If show parsing errors when reading csv /
Si muestra errores cuando lee el csv}

\item{language}{('Español', 'English') Message languages/ Idiomas del mensaje.}

\item{read_format}{\code{"MariaDB"} or \code{"tibble"} establishes how the database
should be stored. In most machines \code{"tibble"} will result in a memory error}

\item{user}{User for \code{dbConnect} i.e. your MariaDB user}

\item{password}{password for \code{dbConnect} i.e. your MariaDB password}

\item{dbname}{Database name for \code{dbConnect} i.e. what database is going to be stored}

\item{host}{Host for \code{dbConnect} i.e. your MariaDB host (usually localhost)}

\item{group}{Group for \code{dbConnect} i.e. your MariaDB group (can be NULL)}

\item{port}{Port connection for MariaDB}

\item{tblname}{Name of table to save in MariaDB}

\item{parse_dictionary}{Change labels to values of dictionary? Default = TRUE
else the data is downloaded and presented as is. / Se agregan etiquetas
a partir del diccionario? Default = TRUE en caso contrario solo se presentan los
datos descargados como estan.}
}
\value{
List of values:
\code{dats} Database table (if MARIADB) or database in tibble (if tibble)
\code{con} Database connection (if MARIADB) or \code{NULL} (if tibble)
\code{dict} List of tibbles containing the whole dictionnary
}
\description{
\code{descarga_datos_abiertos} downloads and labels
all COVID-19 data for Mexico from the Direccion General de EpidemiologIa at
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}.

\code{descarga_datos_abiertos} descarga y etiqueta todos los datos de COVID-19
de Mexico de la Direccion General de EpidemiologIa en
\url{https://www.gob.mx/salud/documentos/datos-abiertos-152127}.
}
\details{
While reading the data it is perfectly normal to get a \code{Warning} for
\code{parsing}: the dataset has issues.

En la lectura de datos es normal tener un \code{Warning} por \code{parsing}:
la base viene chueca.

This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos(language = "Espanol")
}
}
