% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos_variantes_GISAID.R
\encoding{UTF-8}
\name{descarga_datos_variantes_GISAID}
\alias{descarga_datos_variantes_GISAID}
\title{LEE LA BASE DE DATOS DE VARIANTES DE GISAID}
\usage{
descarga_datos_variantes_GISAID(type = c("nacional", "cdmx"))
}
\arguments{
\item{type}{si se desea descargar "nacional" o sólo "cdmx"}
}
\value{
\code{data.frame} con los datos porcentuales y de conteo de variantes
}
\description{
\code{descarga_datos_variantes_GISAID} Lee los datos de variantes publicadas por \code{GISAID}
}
\details{
Cada vez que uses estos datos necesitas citar a \strong{GISAID} así como esta publicación producto
del análisis de dichos datos.

\emph{This is not an official product / este no es un producto oficial}

\strong{Spanish}
Los datos son descargados de manera automatica en mi Github:
\href{https://github.com/RodrigoZepeda/VariantesCovid}{RodrigoZepeda/VariantesCovid}
y esta funcion los lee de ahí
}
\examples{
\dontrun{
variantes_covid <- descarga_datos_variantes_GISAID()
}
}
\references{
Khare, S., et al (2021) GISAID’s Role in Pandemic Response. China CDC Weekly, 3(49): 1049-1051. doi:10.46234/ccdcw2021.255 PMCID: 8668406

Elbe, S. and Buckland-Merrett, G. (2017) Data, disease and diplomacy: GISAID’s innovative contribution to global health. Global Challenges, 1:33-46. doi:10.1002/gch2.1018 PMCID: 31565258

Shu, Y. and McCauley, J. (2017) GISAID: from vision to reality. EuroSurveillance, 22(13) doi:10.2807/1560-7917.ES.2017.22.13.30494 PMCID: PMC5388101
}
