% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_datos_variantes_GISAID.R
\encoding{UTF-8}
\name{descarga_datos_variantes_GISAID}
\alias{descarga_datos_variantes_GISAID}
\title{LEE LA BASE DE DATOS DE VARIANTES DE COVID-19 DE GISAID}
\usage{
descarga_datos_variantes_GISAID(
  nivel = c("nacional", "cdmx"),
  cache = NULL,
  use_cache_on_failure = rlang::is_interactive(),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{nivel}{si se desea descargar \code{"nacional"} o \code{"cdmx"}.}

\item{cache}{cache para \code{pins::board_url}.}

\item{use_cache_on_failure}{parametro para \code{pins::board_url}.}

\item{quiet}{booleana para no imprimir mensajes en la consola.}

\item{...}{parametros adicionales para \code{pins::pin_download}.}
}
\value{
\code{data.frame} con los datos porcentuales y de conteo de variantes
}
\description{
\code{descarga_datos_variantes_GISAID} Lee los datos de variantes del reporte nacional diario en
https://github.com/RodrigoZepeda/VariantesCovid  creado a partir de la informacion de la
Global Initiative on Sharing Avian Influenza Data (GISAID)
}
\details{
Cada vez que uses estos datos necesitas citar a \strong{GISAID} (ver referencias) asi como
el reporte en https://github.com/RodrigoZepeda/VariantesCovid.

Los datos son descargados de manera automatica en mi Github:
\href{https://github.com/RodrigoZepeda/VariantesCovid}{RodrigoZepeda/VariantesCovid} el programa
\code{descarga_datos_variantes_GISAID} se conecta a dicho repositorio, busca si la informacion esta
actualizada y si si la descarga, si no, utiliza informacion almacenada en el \code{cache} local.

La descarga usa el paquete \code{pins}
}
\examples{
\dontrun{
#Descarga de variantes a nivel nacional
variantes_covid <- descarga_datos_variantes_GISAID('nacional')

#Descarga de variantes para CDMX
variantes_covid <- descarga_datos_variantes_GISAID('cdmx')
}
}
\references{
Khare, S., et al (2021) GISAID's Role in Pandemic Response. China CDC Weekly, 3(49):
1049-1051. doi:10.46234/ccdcw2021.255 PMCID: 8668406

Elbe, S. and Buckland-Merrett, G. (2017) Data, disease and diplomacy: GISAID’s
innovative contribution to global health. Global Challenges, 1:33-46.
doi:10.1002/gch2.1018 PMCID: 31565258

Shu, Y. and McCauley, J. (2017) GISAID: from vision to reality. EuroSurveillance,
22(13) doi:10.2807/1560-7917.ES.2017.22.13.30494 PMCID: PMC5388101

Zepeda-Tello, R. (2022). Reporte Nacional de Variantes de COVID-19.
URL: \url{https://github.com/RodrigoZepeda/VariantesCovid}
}
\seealso{
descarga_datos_red_irag descarga_datos_abiertos
}
