% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{estima_rt}
\alias{estima_rt}
\title{RT: Número efectivo de reproducción}
\usage{
estima_rt(
  datos_covid,
  entidades = c("AGUASCALIENTES", "BAJA CALIFORNIA", "BAJA CALIFORNIA SUR", "CAMPECHE",
    "CHIAPAS", "CHIHUAHUA", "CIUDAD DE MÉXICO", "COAHUILA DE ZARAGOZA", "COLIMA",
    "DURANGO", "GUANAJUATO", "GUERRERO", "HIDALGO", "JALISCO", "MÉXICO",
    "MICHOACÁN DE OCAMPO", "MORELOS", "NAYARIT", "NUEVO LEÓN", "OAXACA", "PUEBLA",
    "QUERÉTARO", "QUINTANA ROO", "SAN LUIS POTOSÍ", "SINALOA", "SONORA", "TABASCO",
    "TAMAULIPAS", "TLAXCALA", "VERACRUZ DE IGNACIO DE LA LLAVE", "YUCATÁN", "ZACATECAS"),
  group_by_entidad = TRUE,
  entidad_tipo = c("Unidad Medica", "Residencia", "Nacimiento"),
  fecha_tipo = c("Sintomas", "Ingreso", "Defuncion"),
  tipo_clasificacion = c("Sospechosos", "Confirmados COVID", "Negativo a COVID",
    "Inválido", "No realizado"),
  tipo_paciente = c("AMBULATORIO", "HOSPITALIZADO", "NO ESPECIFICADO"),
  list_name = "estima_rt",
  min_date = as.Date("2021/11/21", format = "\%Y/\%m/\%d"),
  max_date = as.Date(Sys.time()),
  method = "parametric_si",
  config = EpiEstim::make_config(list(mean_si = 3.5, std_si = 1.5)),
  ...
)
}
\arguments{
\item{datos_covid}{If no data is available it automatically downloads COVID-19
information.}

\item{entidades}{Vector con las entidades de las unidades medicas a analizar.
Opciones: \code{AGUASCALIENTES}, \verb{BAJA CALIFORNIA}, \verb{BAJA CALIFORNIA SUR},
\code{CAMPECHE}, \code{CHIAPAS}, \code{CHIHUAHUA}, \verb{CIUDAD DE MEXICO},
\verb{COAHUILA DE ZARAGOZA} , \code{COLIMA}, \code{DURANGO}, \code{GUANAJUATO}, \code{GUERRERO},
\code{HIDALGO}, \code{JALISCO}, \code{MEXICO}, \verb{MICHOACAN DE OCAMPO}, \code{MORELOS},\code{NAYARIT}
\verb{NUEVO LEON}, \code{OAXACA} ,\code{PUEBLA}, \code{QUERETARO},\verb{QUINTANA ROO},
\verb{SAN LUIS POTOSI}, \code{SINALOA}, \code{SONORA}, \code{TABASCO}, \code{TAMAULIPAS},\code{TLAXCALA},
\verb{VERACRUZ DE IGNACIO DE LA LLAVE}, \code{YUCATAN}, \code{ZACATECAS}}

\item{group_by_entidad}{Si junta las entidades en una sola
o bien las muestra por separado sin agrupar.}

\item{entidad_tipo}{Selecciona \verb{Unidad Medica}, \code{Nacimiento} o \code{Residencia}.
por default incluye \verb{Unidad Medica}}

\item{fecha_tipo}{Selecciona \code{Ingreso}, \code{Sintomas} o \code{Defuncion} por default
incluye fecha de \code{Sintomas}}

\item{tipo_clasificacion}{Vector con el tipo de clasificaciones a incluir:
\code{Sospechosos},\verb{Confirmados COVID}, \verb{Negativo a COVID},
\verb{Inv\\u00e1lido}, \verb{No realizado}}

\item{tipo_paciente}{Vector con el tipo de pacientes a incluir. Opciones:
\code{AMBULATORIO}, \code{HOSPITALIZADO}, \verb{NO ESPECIFICADO}}

\item{list_name}{Asigna un nombre en la lista de datos a la base generada}

\item{min_date}{Mínima fecha a partir de la cual estimar el RT}

\item{max_date}{Máxima fecha a partir de la cual estimar el RT}

\item{method}{Method for estimating RT with \code{EpiEstim::estimate_R}}

\item{config}{Configuration for estimating RT with \code{EpiEstim::estimate_R}}

\item{...}{Parámetros adicionales para \code{EpiEstim::estimate_R}.}
}
\value{
Appends a la lista de \code{datos_covid} una nueva entrada de nombre \code{list_name}
(default: \code{casos}) con una base de datos (\code{tibble}) con los
resultados agregados.
\itemize{
\item positividad - Base de datos generara con los datos agregados (el nombre cambia si
se usa \code{list_name}).
\item dict - Diccionario de datos
\item dats - Datos originales (conexion a DB)
\item disconnect  - Funcion para desconectarte de DB
}
}
\description{
\code{estima_rt} Calcula el número efectivo de reproducción por fecha y entidad
}
\details{
Por default calcula el número efectivo de reproducción por estado
This is not an official product / este no es un producto oficial
}
\examples{
\dontrun{
datos_covid <- descarga_datos_abiertos()

# Casos a nivel nacional en los confirmados
datos_covid <- datos_covid |>
  estima_rt(
    tipo_clasificacion = "Confirmados COVID",
    group_by_entidad   = FALSE
  )

# Casos en AGS, CHI en los confirmados
datos_covid |>
  estima_rt(
    entidades = c("CHIHUAHUA", "AGUASCALIENTES"),
    tipo_clasificacion = "Confirmados COVID",
    group_by_entidad = TRUE, 
    list_name = "rt_ch_ags"
  ) |> 
  plot_covid(df_name = "rt_ch_ags", df_date_index = "FECHA_SINTOMAS", 
             df_variable = "Mean(R)", df_covariates = "ENTIDAD_FEDERATIVA")
}

}
