% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lectura.R
\encoding{UTF-8}
\name{read_datos_abiertos}
\alias{read_datos_abiertos}
\alias{read_datos_abiertos_zip}
\alias{read_datos_abiertos_csv}
\alias{read_datos_abiertos_MariaDB}
\title{LEE LA BASE DE DATOS ABIERTOS DE LA DIRECCION GENERAL DE EPIDEMIOLOGIA QUE YA DESCARGASTE}
\usage{
read_datos_abiertos(
  datos_abiertos_path = NULL,
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  read_format = c("MariaDB", "tibble"),
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  ...
)

read_datos_abiertos_zip(
  datos_abiertos_zip_path,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  read_format = c("MariaDB", "tibble"),
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_command = Sys.getenv("unzip_command"),
  unzip_args = Sys.getenv("unzip_args"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  check_unzip_install = TRUE,
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  use_dict = TRUE,
  quiet = FALSE,
  cache_datos = NULL,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name = "datos_abiertos",
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_diccionario_ssa_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

read_datos_abiertos_csv(
  datos_abiertos_unzipped_path,
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  read_format = c("MariaDB", "tibble"),
  driver = RMariaDB::MariaDB(),
  sqlimport = "mysqlimport",
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  tblname = "covidmx",
  nthreads = max(parallel::detectCores() - 1, 1),
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  clear_csv = TRUE,
  quiet = FALSE,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_diccionario_ssa_args = list(),
  prestatement = paste0("SET sql_mode = 'NO_ENGINE_SUBSTITUTION,",
    "NO_AUTO_CREATE_USER';"),
  additional_mysqlimport_flags = c("--local"),
  load_data_infile_options = list(LOW_PRIORITY = "", CONCURRENT = "", LOCAL = "LOCAL",
    PARTITION = ""),
  ...
)

read_datos_abiertos_MariaDB(
  user = Sys.getenv("MariaDB_user"),
  password = Sys.getenv("MariaDB_password"),
  dbname = Sys.getenv("MariaDB_dbname"),
  host = Sys.getenv("MariaDB_host"),
  group = Sys.getenv("MariaDB_group"),
  port = Sys.getenv("MariaDB_port"),
  driver = RMariaDB::MariaDB(),
  tblname = "covidmx",
  diccionario_zip_path = NULL,
  diccionario_unzipped_path = NULL,
  diccionario = NULL,
  download_process = c("pins", "download.file"),
  site.covid.dic = paste0("http://datosabiertos.salud.", "gob.mx/gobmx/salud/datos_a",
    "biertos/diccionario_datos_", "covid19.zip"),
  unzip_args_dict = list(exdir = ".", overwrite = TRUE),
  clear_zip = download_process[1] != "pins",
  clear_csv = TRUE,
  use_dict = TRUE,
  quiet = FALSE,
  use_cache_on_failure = TRUE,
  cache_diccionario = NULL,
  force_download = FALSE,
  show_warnings = TRUE,
  board_url_name_dict = "diccionario_covid",
  download_file_args = list(method = "curl", destfile = tempfile(), quiet = quiet),
  descarga_db_diccionario_ssa_args = list(),
  ...
)
}
\arguments{
\item{datos_abiertos_path}{Camino a los datos abiertos si son un \code{zip} o un \code{csv} o bien
dejar en \code{NULL} si se quiere leer de \code{MariaDB}}

\item{driver}{An \code{SQL} driver for \code{dbConnect}}

\item{sqlimport}{An import command similar to \code{mysqlimport}}

\item{read_format}{\code{"MariaDB"} o \code{"tibble"} establece el formato
de lectura de la base de datos. En la mayoria de los casos \code{"tibble"} va a
resultar en un error de memoria.}

\item{user}{Usuario para \code{dbConnect} i.e. tu usuario de \code{MariaDB}.
Puedes ponerlo como variable ambiental con \code{Sys.setenv("MariaDB_user" = "tu_usuario")}}

\item{password}{password para \code{dbConnect} i.e. tu password de \code{MariaDB}
Puedes ponerlo como variable ambiental con \code{Sys.setenv("MariaDB_password" = "tu_passwodrd")}}

\item{dbname}{Nombre de la base de datos \code{DATABASE} para\code{dbConnect} puedes ponerlo
con \code{Sys.setenv("MariaDB_dbname" = "covidmx")}}

\item{host}{(\strong{opcional}) Host para \code{dbConnect}. Si no sabes cual tienes seguro es \code{localhost}. Puedes
ponerlo con \code{Sys.setenv("MariaDB_host" = "localhost")}}

\item{group}{(\strong{opcional})  Grupo de \code{dbConnect} i.e. tu grupo de \code{MariaDB}. Ponlo como variable ambiental
usando \code{Sys.setenv("MariaDB_group" = "tu_grupo")}}

\item{port}{Puerto de la conexion con \code{MariaDB}. Puedes ponerlo con
\code{Sys.setenv("MariaDB_group" = "tu_grupo")}}

\item{tblname}{Nombre de la tabla dentro de el \code{DATABASE} \code{dbname} donde guardar los datos
por default se llama \code{covidmx}.}

\item{nthreads}{Número de \code{threads} para escribir el archivo en \code{MariaDB}.}

\item{...}{Parametros adicionales para \code{pins::pin_download}}

\item{datos_abiertos_zip_path}{Camino a los datos abiertos si ya los descargaste en \code{zip}}

\item{diccionario_zip_path}{Camino al diccionario si ya losdescargaste en \code{zip}}

\item{diccionario_unzipped_path}{Camino al diccionario \code{csv} si ya lo descargaste y
descomprimiste el archivo \code{zip} en un \code{csv}}

\item{diccionario}{Lo que resulta de realizar una descarga del diccionario
usando \code{descarga_diccionario}}

\item{download_process}{(\strong{opcional})  Metodo para descargar ya sea \code{pins} o \code{download.file}.
Se recomienda \code{pins} pues guarda en memoria la fecha de la ultima descarga y analiza
si ha pasado mas de un dia desde la descarga. En caso afirmativo verifica si el
archivo ha cambiado y si hubo cambios entonces lo descarga.}

\item{site.covid.dic}{Sitio para descarga del diccionario de datos.}

\item{unzip_command}{Forma de extraer la base de datos de datos abiertos. La forma de
llamarla es con \code{system2(unzip_command, args = c(unzip_args, file_download_data))}.}

\item{unzip_args}{Argumentos de extraccion de la base de datos de datos abiertos. La forma de
llamarla es con \code{system2(unzip_command, args = c(unzip_args, file_download_data))}.}

\item{unzip_args_dict}{Lista de argumentos para usar \code{utils::unzip} en el diccionario de datos.}

\item{check_unzip_install}{Bandera de verificacion para checar si tienes lo necesario para
unzippear los datos.}

\item{clear_zip}{Si borrar los archivos \code{.zip} descargados para el diccionario y los datos
abiertos. No se recomienda si estas usando \code{pins}. Ve la nota para mas informacion.}

\item{clear_csv}{Si borrar los archivos \code{.csv} que se generan despues de abrir el zip. El
default es que si.}

\item{use_dict}{Si descargar el diccionario de \code{site.covid.dic}.}

\item{quiet}{(\strong{opcional}) Variable para no mostrar mensajes}

\item{cache_datos}{Direccion donde guardar los datos en memoria usando \code{pins} para no tener
que volver a descargarlos si nada ha cambiado}

\item{use_cache_on_failure}{Booleana. Establece que si no se pueden descargar datos nuevos
utilice los que tenga en memoria.}

\item{cache_diccionario}{Direccion donde guardar el diccionario en memoria usando \code{pins} para no tener
que volver a descargarlo si nada ha cambiado}

\item{force_download}{analiza si cambio el pin y descarga datos nuevos en caso afirmativo}

\item{show_warnings}{si arrojar \code{warnings}}

\item{board_url_name}{Establece el nombre del \code{pins::board_url} para los datos abiertos}

\item{board_url_name_dict}{Establece el nombre del \code{pins::board_url} para los datos abiertos}

\item{download_file_args}{Lista de argumentos adicionales para \code{download.file} de los datos
si se elige este metodo para descargar.}

\item{descarga_db_diccionario_ssa_args}{Lista con argumentos adicionales para el
\code{pins::pin_download} de datos abiertos}

\item{prestatement}{Opciones adicionales para \code{MariaDB} yo pongo las que me han funcionado
en que sea mas veloz la carta.}

\item{additional_mysqlimport_flags}{La instalacion se hace con el siguiente comando para \code{MariaDB}:
\preformatted{
mysqlimport --default-character-set=UTF8 --fields-terminated-by=',' --ignore-lines=1
--fields-enclosed-by='\"' --lines-terminated-by='\\n' --user={user}" --password={password}
--use-threads={nthreads} {silent_flag} {additional_mysqlimport_flags} {dbname} ./{tblname}.csv
}}

\item{load_data_infile_options}{Lista de opciones por si la importacion con \code{mysqlimport} no
funciona se use un \verb{LOAD DATA INFILE} como aquí \url{https://dev.mysql.com/doc/refman/8.0/en/load-data.html}
Las variables se pasan como lista asi por ejemplo si queremos agregar la variable local se
pone \code{list("LOCAL" = "LOCAL")} y si no la queremos \code{list("LOCAL" = "")}.}

\item{datos_abiertos_unzipped_path}{Camino a los datos abiertos \code{csv} si ya los descargaste y
descomprimiste el archivo \code{zip} en un \code{csv}}
}
\value{
List of values:
\itemize{
\item dats - Database table (if MARIADB) or database in tibble (if tibble)
\item con - Database connection (if MARIADB) or \code{NULL} (if tibble)
\item dict - List of tibbles containing the whole dictionary
}
}
\description{
\code{read_datos_abiertos} Lee los datos abiertos almacenados en tu base de \code{MariaDB} que
bajaste con \code{descarga_datos_abiertos}. Intenta de manera automática determinar
si los lee de \code{MariaDB}, \code{csv} ó \code{zip}

La funcion principal es \code{\link[=read_datos_abiertos]{read_datos_abiertos()}} la cual decide si los lee de zip, MARIADB o csv
Tambien puedes usar las auxiliares respectivas
\itemize{
\item \code{\link[=read_datos_abiertos_zip]{read_datos_abiertos_zip()}}     Si sólo descargaste los datos de la DGE en \code{.zip}
\item \code{\link[=read_datos_abiertos_csv]{read_datos_abiertos_csv()}}     Si descargaste los datos de la DGE en \code{.zip} y los descomprimiste.
\item \code{\link[=read_datos_abiertos_MariaDB]{read_datos_abiertos_MariaDB()}} Si ya creaste tu table en \code{MariaDB}
}
}
\examples{
\dontrun{
#Lee los datos de MariaDB una vez descargados
descarga_datos_abiertos()
datos_covid   <- read_datos_abiertos()

#Es lo mismo que:
datos_covid   <- read_datos_abiertos_MariaDB()

#Descarga los datos y lee de un zip
direccion_zip <- descarga_db_datos_abiertos_tbl()
datos_covid   <- read_datos_abiertos(direccion_zip)

#Es lo mismo que:
datos_covid   <- read_datos_abiertos_zip(direccion_zip)

#Descarga los datos y lee de un csv
direccion_zip <- descarga_db_datos_abiertos_tbl()
direccion_csv <- unzip_db_datos_abiertos_tbl(direccion_zip)
datos_covid   <- read_datos_abiertos(direccion_csv)

datos_covid   <- read_datos_abiertos_csv(direccion_csv)
}
}
